// frontend/src/app/core/api/model/review.ts
/**
 * MarketCore API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { User } from './user';


export interface Review { 
    id?: number;
    order_id?: number;
    reviewer_id?: number;
    reviewee_id?: number;
    rating?: number;
    comment?: string;
    created_at?: string;
    reviewer?: User;
}